feat(core): strip display-name prefix from primary universe on migration
Apply the same prefix-stripping treatment to the primary fold's universe
entry as places already receive: when the primary environment has a
captured label, lift the unprefixed body to root.universe.displayName so
the deploy-time prefix system reapplies the bracket cleanly. Without a
label, the universe displayName passes through verbatim.

Extracts the resource-asymmetry warning helpers into a sibling
factorize-environments-warnings module to keep both files within the
project's per-file size cap.
